requirements/development.txt
-r _base.txt
mypy==0.910
pre-commit==2.17.0
pytz==2021.3
pylint==2.12.2
pytest==6.2.4
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-factoryboy==2.1.0